src/backend/line-bot-reply/requirements.txt
line-bot-sdk
flask
python-dotenv
pymongo[srv]